@@ -1,8 +1,8 @@
 package io.github.ititus.nbtviewer.client.util;
 
+import com.mojang.blaze3d.platform.InputConstants;
+import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.settings.KeyBinding;
-import net.minecraft.client.util.InputMappings;
 import net.minecraftforge.client.settings.KeyModifier;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -16,62 +16,59 @@ public final class KeyHelper {
     }
 
     /**
-     * Checks if the given KeyBinding is active, using the bound key, context and modifiers.
+     * Checks if the given KeyMapping is active, using the bound key, context and modifiers.
      *
-     * @param kb KeyBinding
-     * @return true if the KeyBinding is active, otherwise false
+     * @param km KeyMapping
+     * @return true if the KeyMapping is active, otherwise false
      */
-    public static boolean isActive(KeyBinding kb) {
-        if (kb == null) {
+    public static boolean isActive(KeyMapping km) {
+        if (km == null) {
             return false;
-        } else if (kb.isDown()) {
+        } else if (km.isDown()) {
             return true;
-        } else if (kb.isConflictContextAndModifierActive()) {
-            return isKeyPressed(kb);
+        } else if (km.isConflictContextAndModifierActive()) {
+            return isKeyPressed(km);
         }
 
-        return KeyModifier.isKeyCodeModifier(kb.getKey()) && isKeyPressed(kb);
+        return KeyModifier.isKeyCodeModifier(km.getKey()) && isKeyPressed(km);
     }
 
     /**
-     * Checks if the key bound to a KeyBinding is pressed, ignoring context and modifiers.
+     * Checks if the key bound to a KeyMapping is pressed, ignoring context and modifiers.
      *
-     * @param kb KeyBinding
+     * @param km KeyMapping
      * @return true if the key is pressed, otherwise false
      */
-    public static boolean isKeyPressed(KeyBinding kb) {
-        if (kb == null || kb.getKey().getValue() == InputMappings.UNKNOWN.getValue()) {
+    public static boolean isKeyPressed(KeyMapping km) {
+        if (km == null || km.getKey().getValue() == InputConstants.UNKNOWN.getValue()) {
             return false;
         }
 
         try {
             long handle = Minecraft.getInstance().getWindow().getWindow();
-            switch (kb.getKey().getType()) {
-                case KEYSYM:
-                    return InputMappings.isKeyDown(handle, kb.getKey().getValue());
-                case SCANCODE:
-                    return false;
-                case MOUSE:
-                    return GLFW.glfwGetMouseButton(handle, kb.getKey().getValue()) == GLFW.GLFW_PRESS;
-            }
+            return switch (km.getKey().getType()) {
+                case KEYSYM -> InputConstants.isKeyDown(handle, km.getKey().getValue());
+                case SCANCODE -> false;
+                case MOUSE -> GLFW.glfwGetMouseButton(handle, km.getKey().getValue()) == GLFW.GLFW_PRESS;
+            };
         } catch (Exception e) {
-            L.warn("Unable to check keybind status of {}", toString(kb), e);
+            L.warn("Unable to check keybind status of {}", toString(km), e);
         }
 
         return false;
     }
 
-    public static String toString(KeyBinding kb) {
-        if (kb == null) {
+    public static String toString(KeyMapping km) {
+        if (km == null) {
             return "null";
         }
 
-        return kb.getClass().getSimpleName() + "{" +
-                "description=" + kb.getName() +
-                ", key=" + kb.getKey() +
-                ", category=" + kb.getCategory() +
-                ", modifier=" + kb.getKeyModifier() +
-                ", conflictContext=" + kb.getKeyConflictContext() +
+        return km.getClass().getSimpleName() + "{" +
+                "description=" + km.getName() +
+                ", key=" + km.getKey() +
+                ", category=" + km.getCategory() +
+                ", modifier=" + km.getKeyModifier() +
+                ", conflictContext=" + km.getKeyConflictContext() +
                 '}';
     }
 }
